requirements.txt
absl-py==0.11.0
alabaster==0.7.12
appdirs==1.4.4
asn1crypto==0.24.0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
black==20.8b1
bleach==3.1.4
cached-property==1.5.2
cachetools==4.2.1
captum==0.3.1
certifi==2018.1.18
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.3
configparser==5.0.1
cryptography==2.1.4
cycler==0.10.0
dataclasses==0.8
decorator==4.1.2
docker==4.2.0
docker-pycreds==0.4.0
docutils==0.16
flake8==3.9.0
Flask==1.1.2
future==0.18.2
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.25.0
google-auth-oauthlib==0.4.2
grpcio==1.35.0
h5py==3.1.0
httplib2==0.9.2
humanfriendly==8.2
idna==2.6
idna-ssl==1.1.0
imagesize==1.2.0
importlib-metadata==1.6.0
iniconfig==1.1.1
ipython-genutils==0.2.0
isort==5.6.4
itsdangerous==1.1.0
jeepney==0.4.3
Jinja2==2.11.2
joblib==1.0.0
keyring==21.2.0
kiwisolver==1.3.1
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazy-object-proxy==1.4.3
Markdown==3.3.3
markdown-it-py==0.6.2
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mdit-py-plugins==0.2.6
multidict==5.1.0
mypy-extensions==0.4.3
numpy==1.19.5
oauth==1.0.1
oauthlib==3.1.0
packaging==20.9
pathspec==0.8.1
pathtools==0.1.2
pbr==5.5.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.2.0
pkg-resources==0.0.0
pkginfo==1.5.0.1
pluggy==0.13.1
promise==2.3
prompt-toolkit==1.0.15
protobuf==3.14.0
psutil==5.7.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycrypto==2.6.1
pyflakes==2.3.0
pyglet==1.5.0
Pygments==2.6.1
PyICU==1.9.8
pylint==2.6.0
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
pytest==6.2.2
python-dateutil==2.8.1
pytz==2018.3
pyxdg==0.25
PyYAML==5.3.1
readme-renderer==26.0
regex==2021.3.17
reportlab==3.4.0
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.1.5
rope==0.18.0
rsa==4.7
scikit-learn==0.24.1
scipy==1.5.4
SecretStorage==3.1.2
semver==2.9.1
sentry-sdk==0.19.5
shortuuid==1.0.1
simplegeneric==0.8.1
simplejson==3.13.2
six==1.15.0
sklearn==0.0
smmap==3.0.5
snowballstemmer==2.1.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
testresources==2.0.1
threadpoolctl==2.1.0
toml==0.10.1
torch==1.7.0+cu101
torchvision==0.8.1+cu101
tqdm==4.45.0
traitlets==4.3.2
twine==3.1.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
ue4-docker==0.0.62
urllib3==1.26.4
wadllib==1.3.2
wandb==0.10.17
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.1.0
-f https://download.pytorch.org/whl/torch_stable.html
